+++ execute_sql.go
@@ -2,9 +2,13 @@ package main
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
+	"slices"
 	"time"
+
+	"github.com/ngicks/go-iterator-helper/x/exp/xiter"
 
 	"github.com/apstndb/lox"
 
@@ -301,4 +305,75 @@ func executeInformationSchemaBasedStatement(ctx context.Context, session *Sessio
 		Rows:         rows,
 		AffectedRows: len(rows),
 	}, nil
+}
+
+func runAnalyzeQuery(ctx context.Context, session *Session, stmt spanner.Statement, isDML bool) (queryPlan *spannerpb.QueryPlan, commitTimestamp time.Time, metadata *spannerpb.ResultSetMetadata, err error) {
+	if !isDML {
+		queryPlan, metadata, err := session.RunAnalyzeQuery(ctx, stmt)
+		return queryPlan, time.Time{}, metadata, err
+	}
+
+	_, timestamp, queryPlan, metadata, err := runInNewOrExistRwTxForExplain(ctx, session, func() (int64, *spannerpb.QueryPlan, *spannerpb.ResultSetMetadata, error) {
+		plan, metadata, err := session.RunAnalyzeQuery(ctx, stmt)
+		return 0, plan, metadata, err
+	})
+	return queryPlan, timestamp, metadata, err
+}
+
+// runInNewOrExistRwTxForExplain is a helper function for runAnalyzeQuery and ExplainAnalyzeDmlStatement.
+// It execute a function in the current RW transaction or an implicit RW transaction.
+func runInNewOrExistRwTxForExplain(ctx context.Context, session *Session, f func() (affected int64, plan *spannerpb.QueryPlan, metadata *spannerpb.ResultSetMetadata, err error)) (affected int64, ts time.Time, plan *spannerpb.QueryPlan, metadata *spannerpb.ResultSetMetadata, err error) {
+	var implicitRWTx bool
+	if !session.InReadWriteTransaction() {
+		implicitRWTx = true
+		// Start implicit transaction.
+		begin := BeginRwStatement{}
+		if _, err := begin.Execute(ctx, session); err != nil {
+			return 0, time.Time{}, nil, nil, err
+		}
+	}
+
+	affected, plan, metadata, err = f()
+	if err != nil {
+		// once error has happened, escape from implicit transaction
+		rollback := &RollbackStatement{}
+		if _, rollbackErr := rollback.Execute(ctx, session); rollbackErr != nil {
+			err = errors.Join(err, fmt.Errorf("error on rollback: %w", rollbackErr))
+		}
+		return 0, time.Time{}, nil, nil, fmt.Errorf("transaction was aborted: %w", err)
+	}
+
+	var commitTimestamp time.Time
+	if implicitRWTx {
+		// query mode PLAN doesn't have any side effects, but use commit to get commit timestamp.
+		commit := CommitStatement{}
+		txnResult, err := commit.Execute(ctx, session)
+		if err != nil {
+			return 0, time.Time{}, nil, nil, err
+		}
+		commitTimestamp = txnResult.Timestamp
+	}
+
+	return affected, commitTimestamp, plan, metadata, nil
+}
+
+// extractColumnNames extract column names from ResultSetMetadata.RowType.Fields.
+func extractColumnNames(fields []*spannerpb.StructType_Field) []string {
+	return slices.Collect(xiter.Map((*spannerpb.StructType_Field).GetName, slices.Values(fields)))
+}
+
+// parseQueryStats parses spanner.RowIterator.QueryStats.
+func parseQueryStats(stats map[string]any) (QueryStats, error) {
+	var queryStats QueryStats
+
+	b, err := json.Marshal(stats)
+	if err != nil {
+		return queryStats, err
+	}
+
+	err = json.Unmarshal(b, &queryStats)
+	if err != nil {
+		return queryStats, err
+	}
+	return queryStats, nil
 }
